# docs/openapi/autoflow-api.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AutoFlow API",
    "version": "v1",
    "description": "Contrato OpenAPI versionado da API AutoFlow. Exportado a partir do contrato SpringDoc da aplicação e complementado com parâmetros de execução, autenticação e exemplos em docs/openapi/README.md."
  },
  "servers": [
    {
      "url": "http://localhost:8081",
      "description": "API local (backend)"
    },
    {
      "url": "http://localhost:4200/api",
      "description": "API local pelo proxy do frontend"
    }
  ],
...
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/servicos/{id}": {
      "get": {
        "operationId": "consultarServico",
        "summary": "Consultar serviço",
        "tags": [
          "serviços"
        ],
        "description": "Consultar serviço Permissão: ADMIN, ATENDENTE ou MECANICO.",
        "responses": {
          "200": {
            "description": "Operação concluída com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServicoResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ]
      },
      "delete": {
        "operationId": "inativarServico",
        "summary": "Inativar serviço",
        "tags": [
          "serviços"
        ],
        "description": "Inativar serviço Permissão: ADMIN.",
        "responses": {
          "200": {
            "description": "Operação concluída com sucesso.",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string",
                  "example": "Serviço inativado com sucesso"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ]
      }
    },
...
    "/servicos/metricas/tempo-medio": {
      "get": {
        "operationId": "listarTempoMedioServicos",
        "summary": "Consultar tempo médio por serviço",
        "tags": [
          "serviços"
        ],
        "description": "Consultar tempo médio por serviço Permissão: ADMIN.",
        "responses": {
          "200": {
            "description": "Operação concluída com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TempoMedioServicoResponse"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
...
    "/ordens-servico": {
      "post": {
        "operationId": "criarOrdemServico",
        "summary": "Abrir ordem de serviço",
        "tags": [
          "ordens de serviço"
        ],
        "description": "Abrir ordem de serviço Permissão: ADMIN ou ATENDENTE.",
        "responses": {
          "201": {
            "description": "Operação concluída com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrdemServicoResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CriarOrdemServicoRequest"
              }
            }
          }
        }
      },
      "get": {
        "operationId": "listarOrdensServico",
        "summary": "Listar ordens de serviço",
        "tags": [
          "ordens de serviço"
        ],
        "description": "Listar ordens de serviço Permissão: ADMIN, ATENDENTE ou MECANICO.",
        "responses": {
          "200": {
            "description": "Operação concluída com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedModelOrdemServicoResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        },
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "minimum": 0
            },
            "example": 0
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10,
              "minimum": 1
            },
            "example": 10
          },
          {
            "name": "cliente",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "numeroOs",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "RECEBIDA",
                "EM_DIAGNOSTICO",
                "AGUARDANDO_APROVACAO",
                "EM_EXECUCAO",
                "FINALIZADA",
                "ENTREGUE"
              ]
            },
            "example": "RECEBIDA"
          }
        ]
      }
    },
    "/ordens-servico/{numeroOs}": {
      "get": {
        "operationId": "detalharOrdemServico",
        "summary": "Detalhar ordem de serviço",
        "tags": [
          "ordens de serviço"
        ],
        "description": "Detalhar ordem de serviço Permissão: ADMIN, ATENDENTE ou MECANICO.",
        "responses": {
          "200": {
            "description": "Operação concluída com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrdemServicoDetalheResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "parameters": [
          {
            "name": "numeroOs",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "OS-2026-0001"
          }
        ]
      }
    },
    "/ordens-servico/{numeroOs}/status": {
      "get": {
        "operationId": "consultarStatusOrdemServico",
        "summary": "Consultar status enxuto da OS",
        "tags": [
          "ordens de serviço"
        ],
        "description": "Consultar status enxuto da OS Permissão: ADMIN, ATENDENTE, MECANICO ou CLIENTE.",
        "responses": {
          "200": {
            "description": "Operação concluída com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusOrdemServicoResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "parameters": [
          {
            "name": "numeroOs",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "OS-2026-0001"
          }
        ]
      }
    },
    "/ordens-servico/metricas/tempo-medio": {
      "get": {
        "operationId": "calcularTempoMedioOrdensServico",
        "summary": "Consultar tempo médio de OS finalizadas",
        "tags": [
          "ordens de serviço"
        ],
        "description": "Consultar tempo médio de OS finalizadas Permissão: ADMIN.",
        "responses": {
          "200": {
            "description": "Operação concluída com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TempoMedioOrdemServicoResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/ordens-servico/{numeroOs}/servicos": {
      "post": {
        "operationId": "incluirServicosOrdemServico",
        "summary": "Incluir serviços na OS",
        "tags": [
          "ordens de serviço"
        ],
        "description": "Incluir serviços na OS Permissão: ADMIN, ATENDENTE ou MECANICO.",
        "responses": {
          "202": {
            "description": "Operação concluída com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrdemServicoResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "parameters": [
          {
            "name": "numeroOs",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "OS-2026-0001"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/ServicoSolicitadoRequest"
                }
              }
            }
          }
        }
      }
    },
    "/ordens-servico/{numeroOs}/mecanico": {
      "patch": {
        "operationId": "atribuirMecanico",
        "summary": "Atribuir mecânico à OS",
        "tags": [
          "ordens de serviço"
        ],
        "description": "Atribuir mecânico à OS Permissão: ADMIN ou ATENDENTE.",
        "responses": {
          "202": {
            "description": "Operação concluída com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrdemServicoResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "parameters": [
          {
            "name": "numeroOs",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "OS-2026-0001"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IncluirMecanicoRequest"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Use o token retornado por POST /auth/login no header Authorization: Bearer <token>."
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Requisição inválida, JSON inválido, validação ou regra de negócio.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "erro": "JSON inválido"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "JWT ausente/inválido ou token público ausente, inválido ou expirado.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "erro": "Token inválido ou expirado"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Usuário autenticado sem a role necessária.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "erro": "Acesso negado"
            }
          }
        }
      },
      "NotFound": {
        "description": "Recurso não encontrado.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "erro": "Recurso não encontrado"
            }
          }
        }
      },
...
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "erro": {
            "type": "string",
            "example": "Recurso não encontrado"
          },
          "campos": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          }
        }
      },
...
      "ServicoResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "nome": {
            "type": "string",
            "example": "Troca de óleo"
          },
          "descricao": {
            "type": "string",
            "example": "Troca de óleo e filtro"
          },
          "valor": {
            "type": "number",
            "format": "double",
            "example": 150
          },
          "ativo": {
            "type": "boolean",
            "example": true
          }
        }
      },
...
      "ServicoSolicitadoRequest": {
        "type": "object",
        "properties": {
          "servicoId": {
            "type": "integer",
            "format": "int64",
            "example": 1
          }
        },
        "required": [
          "servicoId"
        ]
      },
      "VeiculoOrdemServicoRequest": {
        "type": "object",
        "properties": {
          "placa": {
            "type": "string",
            "example": "ABC1D23"
          },
          "marca": {
            "type": "string",
            "example": "Toyota"
          },
          "modelo": {
            "type": "string",
            "example": "Corolla"
          },
          "ano": {
            "type": "integer",
            "example": 2020
          }
        },
        "required": [
          "placa"
        ]
      },
      "CriarOrdemServicoRequest": {
        "type": "object",
        "properties": {
          "cpfCnpj": {
            "type": "string",
            "example": "12345678901"
          },
          "veiculo": {
            "$ref": "#/components/schemas/VeiculoOrdemServicoRequest"
          },
          "servicosSolicitados": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ServicoSolicitadoRequest"
            }
          }
        },
        "required": [
          "cpfCnpj",
          "veiculo",
          "servicosSolicitados"
        ]
      },
      "IncluirMecanicoRequest": {
        "type": "object",
        "properties": {
          "mecanicoId": {
            "type": "integer",
            "format": "int64",
            "example": 2
          },
          "mecanicoEmail": {
            "type": "string",
            "format": "email",
            "example": "mecanico1@autoflow.com"
          }
        }
      },
...
      "OrdemServicoResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "numeroOs": {
            "type": "string",
            "example": "OS-2026-0001"
          },
          "clienteNome": {
            "type": "string",
            "example": "Maria da Silva"
          },
          "clienteCpfCnpj": {
            "type": "string",
            "example": "12345678901"
          },
          "status": {
            "type": "string",
            "enum": [
              "RECEBIDA",
              "EM_DIAGNOSTICO",
              "AGUARDANDO_APROVACAO",
              "EM_EXECUCAO",
              "FINALIZADA",
              "ENTREGUE"
            ],
            "example": "RECEBIDA"
          },
          "dataAbertura": {
            "type": "string",
            "format": "date-time",
            "example": "2026-08-23T14:30:00"
          },
          "execucaoIniciadaEm": {
            "type": "string",
            "format": "date-time",
            "example": "2026-08-23T14:30:00"
          },
          "finalizadaEm": {
            "type": "string",
            "format": "date-time",
            "example": "2026-08-23T14:30:00"
          },
          "entregueEm": {
            "type": "string",
            "format": "date-time",
            "example": "2026-08-23T14:30:00"
          },
          "servicos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ServicoOsResponse"
            }
          },
          "acompanhamentoUrl": {
            "type": "string",
            "format": "uri",
            "example": "http://localhost:4200/public/ordens-servico/acompanhamento?token=<token>"
          }
        }
      },
...
      "OrdemServicoDetalheResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "numeroOs": {
            "type": "string",
            "example": "OS-2026-0001"
          },
          "status": {
            "type": "string",
            "enum": [
              "RECEBIDA",
              "EM_DIAGNOSTICO",
              "AGUARDANDO_APROVACAO",
              "EM_EXECUCAO",
              "FINALIZADA",
              "ENTREGUE"
            ],
            "example": "RECEBIDA"
          },
          "dataAbertura": {
            "type": "string",
            "format": "date-time",
            "example": "2026-08-23T14:30:00"
          },
          "ultimaAtualizacao": {
            "type": "string",
            "format": "date-time",
            "example": "2026-08-23T14:30:00"
          },
          "execucaoIniciadaEm": {
            "type": "string",
            "format": "date-time",
            "example": "2026-08-23T14:30:00"
          },
          "finalizadaEm": {
            "type": "string",
            "format": "date-time",
            "example": "2026-08-23T14:30:00"
          },
          "entregueEm": {
            "type": "string",
            "format": "date-time",
            "example": "2026-08-23T14:30:00"
          },
          "cliente": {
            "$ref": "#/components/schemas/ClienteOrdemServicoResponse"
          },
          "veiculo": {
            "$ref": "#/components/schemas/VeiculoOrdemServicoResponse"
          },
          "servicos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ServicoOsResponse"
            }
          },
          "orcamentoAtual": {
            "$ref": "#/components/schemas/OrcamentoResumoResponse"
          },
          "diagnostico": {
            "$ref": "#/components/schemas/DiagnosticoDetalheResponse"
          }
        }
      },
      "StatusOrdemServicoResponse": {
        "type": "object",
        "properties": {
          "numeroOs": {
            "type": "string",
            "example": "OS-2026-0001"
          },
          "status": {
            "type": "string",
            "enum": [
              "RECEBIDA",
              "EM_DIAGNOSTICO",
              "AGUARDANDO_APROVACAO",
              "EM_EXECUCAO",
              "FINALIZADA",
              "ENTREGUE"
            ],
            "example": "RECEBIDA"
          },
          "ultimaAtualizacao": {
            "type": "string",
            "format": "date-time",
            "example": "2026-08-23T14:30:00"
          }
        }
      },
      "TempoMedioOrdemServicoResponse": {
        "type": "object",
        "properties": {
          "quantidadeOrdensFinalizadas": {
            "type": "integer",
            "format": "int64",
            "example": 10
          },
          "tempoMedioSegundos": {
            "type": "number",
            "format": "double",
            "example": 3600
          },
          "tempoMedioMinutos": {
            "type": "number",
            "format": "double",
            "example": 60
          },
          "tempoMedioHoras": {
            "type": "number",
            "format": "double",
            "example": 1
          }
        }
      },
...
      "TempoMedioServicoResponse": {
        "type": "object",
        "properties": {
          "servicoId": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "nomeServico": {
            "type": "string",
            "example": "Troca de óleo"
          },
          "quantidadeExecucoes": {
            "type": "integer",
            "format": "int64",
            "example": 10
          },
          "tempoMedioSegundos": {
            "type": "number",
            "format": "double",
            "example": 3600
          },
          "tempoMedioMinutos": {
            "type": "number",
            "format": "double",
            "example": 60
          },
          "tempoMedioHoras": {
            "type": "number",
            "format": "double",
            "example": 1
          }
        }
      },
...
      "PagedModelOrdemServicoResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrdemServicoResponse"
            }
          },
          "pageable": {
            "type": "object"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "totalPages": {
            "type": "integer",
            "example": 1
          },
          "size": {
            "type": "integer",
            "example": 10
          },
          "number": {
            "type": "integer",
            "example": 0
          },
          "first": {
            "type": "boolean",
            "example": true
          },
          "last": {
            "type": "boolean",
            "example": true
          },
          "numberOfElements": {
            "type": "integer",
            "example": 1
          },
          "empty": {
            "type": "boolean",
            "example": false
          }
        }
      },
...
    }
  }
}